+++ lib/byebug/commands/method.rb
@@ -3,16 +3,12 @@ module Byebug
     include Columnize
 
     def regexp
-      /^\s* m(?:ethod)? \s+ ((iv)|(i(:?nstance)?)\s+)?/x
+      /^\s* m(?:ethod)? \s+ (i(:?nstance)?\s+)?/x
     end
 
     def execute
       obj = bb_eval(@match.post_match)
-      if @match[1] == 'iv'
-        obj.instance_variables.sort.each do |v|
-          print "#{v} = #{obj.instance_variable_get(v).inspect}\n"
-        end
-      elsif @match[1]
+      if @match[1]
         print "#{columnize(obj.methods.sort(), Setting[:width])}\n"
       elsif !obj.kind_of?(Module)
         print "Should be Class/Module: #{@match.post_match}\n"
@@ -28,7 +24,6 @@ module Byebug
 
       def description
         %{m[ethod] i[nstance] <obj>\tshow methods of object
-          m[ethod] iv <obj>\t\tshow instance variables of object
           m[ethod] <class|module>\t\tshow instance methods of class or module}
       end
     end